chore: remove service worker

+++ packages/chat-server/scripts/main.ts
@@ -10,10 +10,12 @@ await $`pnpm --prefix ../chat build`
 await fs.copy(path.resolve(__dirname, '../../chat/dist'), path.resolve(__dirname, '../dist/public'))
 await $`docker buildx build --platform=linux/amd64 -t chat-server .`
 
-await $`docker stop $(docker ps -aq)`
-await $`docker rm $(docker ps -aq)`
+if ((await $`docker ps -aq`).stdout.trim().length > 0) {
+  await $`docker stop $(docker ps -aq)`
+  await $`docker rm $(docker ps -aq)`
+}
 const e = Object.entries(c)
   .map(([k, v]) => `-e "${k}=${v}"`)
   .join(' ')
 await $`docker run ${e} -dp 8090:8080 chat-server`
-console.log(`docker run ${e} -dp 8090:8080 chat-server`)
+console.log(`run in http://localhost:8090`)
